requirements.txt
pika
django
Camper-Logger-Exception==0.1.3
django-crispy-forms